Merge updates from ocm-plus repo (#13)
* Update operator sdk (#61)

* Remove vendor directory

As far as I could tell, this wasn't being used, and it wasn't updated
or checked by anything in the make targets.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update operator-sdk to 1.13.1

The newer scaffold for using kubebuilder changes the location of several
files:
 - Controller code moves from ./pkg/controllers/* to ./controllers/*
 - The manager code moves from ./cmd/manager/main.go to ./main.go

Usually, the deployment YAML would be a `kustomize` target built from
a ./config/ tree. In this case, to keep things simpler, there are only a
few kustomize items under ./deploy/, which build into the
./deploy/operator.yaml. This keeps the deployment process almost
unchanged.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/15551

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Enable leader election by default

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove references to cmd/manager

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Use the latest tag instead of latest-dev (#62)

The latest-dev images are old and no longer used. We are pushing new
changes to latest instead.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Fix tests (#67)

* Use a blocking kubectl function for tests

The normal `Kubectl` test function does not wait for the command to
return, which means it can fail silently, or cause concurrency problems.
The KubectlWithOutput command waits for the command to complete, and we
can just discard the output and check the error.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove incorrect test

This test ensured that a Policy could be created without a spec. But the
Policy CRD now requires a `spec`.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Allow using the legacy leader election method (#66)

Signed-off-by: Will Kutler <wkutler@redhat.com>

* Fix some upstream coordinates

The image tag used in this repo is `edge`, not `latest`.

The makefile should install the CRDs from the `-io` propagator repo,
instead of the other one.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

Co-authored-by: William Kutler <57921170+willkutler@users.noreply.github.com>